requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.16.0
cryptography==41.0.5
distlib==0.3.6
Django==4.2.6
django-cors-headers==4.3.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-writable-nested==0.7.0
filelock==3.10.7
mysqlclient==2.2.0
Pillow==10.1.0
platformdirs==3.2.0
psycopg==3.1.12
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vboxapi==1.0
virtualenv==20.21.0
